% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkNCDB.R
\name{mkNCDB}
\alias{mkNCDB}
\title{Make NCDB R Binary}
\usage{
mkNCDB(df, ncdbHome = "~/data/NCDB", outFile = "cancDef")
}
\arguments{
\item{df}{Tibble returned by pickFields}

\item{ncdbHome}{Location of NCDB fixed width format ASCII files.}

\item{outFile}{Name of output file, which will placed in \code{ncdbHome}.}
}
\value{
Tibble (big data file) saved as canc in \code{outFile}.
}
\description{
Uses output of \code{pickFields()} to produce an R binary of the NCDB data.
}
\note{
Inspired by \code{mkSEER()} in \pkg{SEERaBomb}.
}
\examples{
\dontrun{
library(NCDBR)
d=getFields()
d=pickFields(d)
canc=mkNCDB(d) }
}
