% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickFields.R
\name{pickFields}
\alias{pickFields}
\title{Picks Fields of Focus}
\usage{
pickFields(d, picks = c("casenum", "facID", "fac", "facLoc", "agedx",
  "sex", "race", "ins", "inc", "educ", "urban", "crow", "charlson",
  "seqnum", "CoC", "yrdx", "histo3", "stageC", "d2t", "radiatn", "d2c",
  "chemo", "hct", "surv", "alive"))
}
\arguments{
\item{d}{Tibble produced by \code{getFields()}.}

\item{picks}{Short names of fields you wish to keep.}
}
\value{
Input dataframe with rows not wanted collapsed to strings. This output feeds into \code{mkNCDB()}.
}
\description{
Takes the output of \code{getField()} and squashes down rows that are not of interest to prepare the 
tibble for use as the input to \code{mkNCDB()}.
}
\note{
Inspired by a function of the same name in \pkg{SEERaBomb}.
}
\examples{
\dontrun{
library(NCDBR)
d=getFields()
pickFields(d) }
}
