% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFields.R
\name{getFields}
\alias{getFields}
\title{Get Field Info in SAS File}
\usage{
getFields(year = 2016)
}
\arguments{
\item{year}{year of NCDB data files (only 2016 supported)}
}
\value{
Tibble with short names (\code{nms}, which contain lower case letters if shortened), 
medium-sized names (\code{names}, all caps, default in .sas file), long names [\code{labs}, which get mapped to 
column object attributes in the output of \code{mkNCDB()}], and a list column \code{Levs} where factor rows contain tibbles  
whose n column holds factor numbers and val column holds string labels to which they map.
}
\description{
The NCDB data comes with a .sas file that helps SAS users. Here \code{getFields()} parses that 
.sas file to create a tibble with information about field start and stop postions and factor levels.
This tibble is then the input into \code{pickFields()}.
}
\note{
The 2013 sas file needed manual editing to remove foreign characters at lines
548, 660-662, and  698.  Use library(tools); showNonASCIIfile() to nonascii of 2016 file.
Fixed versions are in the \file{extdata} folder.
Inspired by a function of the same name in \pkg{SEERaBomb}.
}
\examples{
\dontrun{
library(NCDBR) 
getFields() }
}
